requirements.txt
Click==7.0
docxtpl==0.6.3
Flask==1.1.1
Flask-SQLAlchemy==2.4.0
gunicorn==20.0.0
itsdangerous==1.1.0
Jinja2==2.10.1
lxml==4.4.0
MarkupSafe==1.1.1
psycopg2==2.8.3
python-docx==0.8.7
six==1.12.0
SQLAlchemy==1.3.8
Werkzeug==0.15.5
google-api-python-client==1.7.11
google-auth==1.8.2
google-auth-httplib2==0.0.3
